requirements.txt
aiobotocore==2.5.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
botocore==1.29.76
cashews==6.2.0
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-profiler==1.2.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
motor==3.1.2
multidict==6.0.4
netaddr~=0.8.0
pydantic==1.10.9
pymongo==4.3.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
wrapt==1.15.0
yarl==1.9.2
uvicorn==0.22.0
